+++ rust/wxdragon-sys/cpp/include/widgets/wxd_combobox.h
@@ -13,18 +13,23 @@ WXD_EXPORTED void
 wxd_ComboBox_Clear(wxd_ComboBox_t* combo);
 WXD_EXPORTED int
 wxd_ComboBox_GetSelection(wxd_ComboBox_t* combo);
+
 WXD_EXPORTED int
-wxd_ComboBox_GetStringSelection(wxd_ComboBox_t* combo, char* buffer, int buffer_len);
+wxd_ComboBox_GetStringSelection(wxd_ComboBox_t* combo, char* buffer, size_t buffer_len);
+
 WXD_EXPORTED void
 wxd_ComboBox_SetSelection(wxd_ComboBox_t* combo, int index);
+
 WXD_EXPORTED int
-wxd_ComboBox_GetString(wxd_ComboBox_t* combo, int index, char* buffer, int buffer_len);
+wxd_ComboBox_GetString(wxd_ComboBox_t* combo, int index, char* buffer, size_t buffer_len);
+
 WXD_EXPORTED unsigned int
 wxd_ComboBox_GetCount(wxd_ComboBox_t* combo);
 WXD_EXPORTED void
 wxd_ComboBox_SetValue(wxd_ComboBox_t* combo, const char* value);
+
 WXD_EXPORTED int
-wxd_ComboBox_GetValue(wxd_ComboBox_t* combo, char* buffer, int buffer_len);
+wxd_ComboBox_GetValue(wxd_ComboBox_t* combo, char* buffer, size_t buffer_len);
 
 // Text Selection Functions (inherited from wxTextEntry)
 WXD_EXPORTED void
